Доработка страницы добавления учителя

+++ FRDZ School Web/Models/Lesson.cs
@@ -9,14 +9,14 @@ namespace FRDZ_School_Web.Models
         [Key]
         [Column("Lesson_Num")]
         [DisplayName("Номер")]
-        public decimal Id { get; set; }
+        public int Id { get; set; }
 
         [DisplayName("Описание")]
         public string? Lesson_Description { get; set; }
 
         [DisplayName("Предмет")]
         [Required]
-        public decimal Code_Obj { get; set; }
+        public int Code_Obj { get; set; }
 
         [DisplayName("Дата проведения")]
         [Required]
@@ -24,7 +24,7 @@ namespace FRDZ_School_Web.Models
 
         [DisplayName("Учитель")]
         [Required]
-        public decimal Teacher_Num { get; set; }
+        public int Teacher_Num { get; set; }
 
         [DisplayName("Домашнее задание")]
         public string? Home_Task { get; set; }
